+++ src/main/java/com/ordertogether/team14_be/order/details/service/SmsService.java
@@ -1,7 +1,7 @@
 package com.ordertogether.team14_be.order.details.service;
 
 import com.ordertogether.team14_be.member.persistence.entity.Member;
-import com.ordertogether.team14_be.order.details.sms.dto.SmsReq;
+import com.ordertogether.team14_be.order.details.dto.get.SmsReq;
 import com.ordertogether.team14_be.order.details.util.SmsUtil;
 import com.ordertogether.team14_be.spot.entity.Spot;
 import com.ordertogether.team14_be.spot.repository.SimpleSpotRepository;
